+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.10.0]
+ ### Added
+* Revived option to pre-fill template cache (`ehrbase.cache.template-init-on-startup`, default: `false`) ([#1394](https://github.com/ehrbase/ehrbase/pull/1394))
+* Added experimental `AQL` support for `FOLDER` (`ehrbase.aql.experimental.aqlOnFolder`, default: `false`) ([#1401](https://github.com/ehrbase/ehrbase/pull/1401))
+ ### Changed
+* Feature toggle for ehrscape API (`ehrbase.rest.ehrscape.enabled`, default: `false`) ([#1415](https://github.com/ehrbase/ehrbase/pull/1415))
+ ### Fixed
+* Validate that compositions only contain nodes that are defined by the template (`ehrbase.validation.checkForExtraNodes`, default: `true`)  ([#1424](https://github.com/ehrbase/ehrbase/pull/1424))
+
 ## [2.9.0]
  ### Added
  ### Changed 
@@ -109,3 +118,4 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 [2.8.0]: https://github.com/ehrbase/ehrbase/compare/v2.7.0...v2.8.0
 [2.8.1]: https://github.com/ehrbase/ehrbase/compare/v2.8.0...v2.8.1
 [2.9.0]: https://github.com/ehrbase/ehrbase/compare/v2.8.1...v2.9.0
+[2.10.0]: https://github.com/ehrbase/ehrbase/compare/v2.9.0...v2.10.0